added functionality which loads .bashrc from the users home directory
