+++ src/ssb_project_cli/ssb_project/util.py
@@ -98,3 +98,28 @@ def execute_command(
         print(success_desc)
 
     return result
+
+
+def get_kernels_dict() -> dict[str, str]:
+    """Makes a dictionary of installed kernel specifications.
+
+    Returns:
+        kernel_dict: Dictionary of installed kernel specifications
+    """
+    kernels_process = subprocess.run(  # noqa S607
+        ["jupyter", "kernelspec", "list"], capture_output=True
+    )
+    kernels_str = ""
+    if kernels_process.returncode == 0:
+        kernels_str = kernels_process.stdout.decode("utf-8")
+    else:
+        print("An error occured while looking for installed kernels.")
+        exit(1)
+    kernel_dict = {}
+    for kernel in kernels_str.split("\n")[1:]:
+        line = " ".join(kernel.strip().split())
+        line = line.replace("%s ", "").strip()
+        if len(line.split(" ")) == 2:
+            k, v = line.split(" ")
+            kernel_dict[k] = v
+    return kernel_dict
